Introducing com.sun.jump.executive.JUMPApplicationProxy APIs.
JUMPApplicationProxy is a proxy to control JUMPApplication's lifecycle in the isolate as well as providing a way to associate JUMPApplication with a JUMPIsolateProxy pointing to the isolate it's running in in the executive.  The instance(s) of JUMPApplicationProxy(ies) are returned from JUMPIsolateProxy.startApp.  The lifecycle methods previously defined in JUMPIsolateProxy, such as pauseApp and destoryApp, are now moved to the JUMPApplicationProxy.

Reviewed by Nedim and Brandon.