Reviewed by Nedim Fresko.
Changes:

Added a new MIDlet container factory. The container is in the MIDP repository. The container is a prototype that will start a MIDlet when it is selected to be launched with the simple JUMP graphical application manager and exit the process when the MIDlet is destroyed.

Added more JavaDoc for JUMPAppContainer class.

Fixed the JUMPIsolateProcess to exit and not hang when initialation fails.

Fixed the JUMPIsolateProcess to send back a failure when the app is not destroyed in the case where destroyApp(force=false) is called.